% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coordinateDescent_naive}
\alias{coordinateDescent_naive}
\title{TITLE OF THIS FUNCTION}
\usage{
coordinateDescent_naive(X, Y, lambda, precision, max_iter)
}
\arguments{
\item{X}{gives the arma::mat matrix of regressors in this partition of all regressors (there are m partitions
in total, and we run coordinate descent on each partition)}

\item{Y}{gives the arma::vec vector of observations that we project X onto}

\item{lambda}{gives the double giving our lambda coefficient}

\item{precision}{gives the convergence criterion (how close two subsequent iterations should be before termination)}

\item{max_iter}{gives the maximum number of iterations in the inner update loop of coordinate descent}
}
\value{
an updated version of beta
        NEEDS CHANGING:
        - we should compute Xi'Xi once for each i, and then supply it as another parameter?
        - we should be able to just return a pointer to a (changed) beta (more efficient than local copies!)
}
\description{
TITLE OF THIS FUNCTION
}

