% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxFunctions.R
\name{lassoCoefParallel}
\alias{lassoCoefParallel}
\title{Lasso coefficients (parallel implementation)}
\usage{
lassoCoefParallel(data, Y, nlambda = 1, lambda, intercept = FALSE, m)
}
\arguments{
\item{data}{is a list of m matrices X. Each X gives the nxp matrix of regressors,
each column corresponding to a different regressor}

\item{Y}{gives the nx1 vector of observations we wish to approximate with a linear model of type Y = Xb + e}

\item{nlambda}{how many lambda should be tested. Right now only \code{nlambda=1} is supported.}

\item{lambda}{gives the (fixed) penalty magnitude in the LASSO fit of the algorithm}

\item{intercept}{determines whether to include an intercept in the model or not}
}
\value{
The coefficients of the Lasso regression.
}
\description{
Lasso coefficients (parallel implementation)
}
\details{
This function uses glmnet package to compute Lasso coefficient in a parallel way.

Check \code{lassoCoef} for a not parallelized version of this function.
}
\author{
Samuel Davenport, Jack Carter, Giulio Morina, Jeremias Knoblauch
}

