% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squareRootSymmetric}
\alias{squareRootSymmetric}
\title{Square root of a symmetric matrix}
\usage{
squareRootSymmetric(M)
}
\arguments{
\item{M}{a symmetric matrix}
}
\value{
A matrix C which is the square root matrix of M (i.e. C*C=M)
}
\description{
Square root of a symmetric matrix
}
\details{
To compute the square root of the matrix, \code{sqrtmat_sympd} function of Armadillo library is used.
}
\note{
This function is about 100 times faster than R function \code{sqrtm} (contained in expm package).

Note that no check is done to test if the matrix M is actually symmetric.
}
\examples{
require(expm)
require(rbenchmark)
A <- matrix(rnorm(10000,mean=10,sd=5),nrow=100)
A_symm <- A\%*\%t(A)
benchmark(sqrtm(A_symm),squareRootSymmetric(A_symm), order='relative')

}
\author{
Samuel Davenport, Jack Carter, Giulio Morina, Jeremias Knoblauch
}

