% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{standardizeMatrix}
\alias{standardizeMatrix}
\title{Standardize a matrix so that its mean is equal to 0}
\usage{
standardizeMatrix(M)
}
\arguments{
\item{M}{a matrix}
}
\value{
A matrix with mean 0
}
\description{
Standardize a matrix so that its mean is equal to 0
}
\details{
The matrix is standardized by subtracting each column with the mean of that column.
}
\note{
In general, this function does not return a matrix with variance equal to 1.

This function is about 2.5x times faster than doing \code{scale(M,scale=FALSE)}.
}
\examples{
require(rbenchmark)
M <- matrix(rnorm(1000*5000,10,5), nrow=1000)
benchmark(scale(M,scale=FALSE),standardizeMatrix(M), order='relative')

}

