% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{standardizeVector}
\alias{standardizeVector}
\title{Standardize a vector}
\usage{
standardizeVector(V)
}
\arguments{
\item{V}{a vector}
}
\value{
Returns a vector with mean equal to 0.
}
\description{
Standardize a vector
}
\details{
To compute the standardized vector, each of its entries is subtracted with the vector's mean.
}
\note{
In general, this function does not return a vector with variance equal to 1.

This function is about 2x slower than directly computing \code{v-mean(v)}, but it is faster than doing
\code{scale(v,scale=FALSE)}
}
\examples{
require(rbenchmark)
v <- 1:5000000
benchmark({v-mean(v)},standardizeVector(v),scale(v,scale=FALSE), order='relative')

}
\author{
Samuel Davenport, Jack Carter, Giulio Morina, Jeremias Knoblauch
}

