% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_naive}
\alias{update_naive}
\title{TITLE OF THIS FUNCTION}
\usage{
update_naive(beta, X, Y, lambda, n, p)
}
\arguments{
\item{beta}{gives the arma::vec vector of beta-estimates at the current iteration of Coordinate descent}

\item{X}{gives the arma::mat matrix of regressors in this partition of all regressors (there are m partitions
in total, and we run coordinate descent on each partition)}

\item{Y}{gives the arma::vec vector of observations that we project X onto}

\item{lambda}{gives the double giving our lambda coefficient}

\item{n}{gives an integer corresponding to the number of rows (observations) of X (Y)}

\item{p}{gives the integer corresponding to the number of columns/regressors contained in X}
}
\value{
an updated version of beta
        NEEDS CHANGING:
        - we should compute Xi'Xi once for each i, and then supply it as another parameter?
        - we should be able to just return a pointer to a (changed) beta (more efficient than local copies!)
}
\description{
TITLE OF THIS FUNCTION
}

